backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.12
pydantic==2.9.0
python-dotenv==1.0.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database / storage
supabase==2.9.1

# AI providers
anthropic==0.39.0
openai==1.51.0
tiktoken==0.12.0

# HTTP (Catastro WFS async calls)
httpx==0.27.2

# PDF extraction — PyMuPDF is primary, pypdf is fallback
PyMuPDF==1.27.2.2
pypdf==5.0.1

# OCR for scanned PDFs (the Tesseract binary itself is installed
# via apt in the Dockerfile — this is just the Python wrapper)
pytesseract==0.3.13
pillow==12.2.0

# DOCX generation (memoria, anejos, pliegos)
python-docx==1.1.2
